+++ public/ide/modules/ide/js/parts/nav/renderService.js
@@ -504,6 +504,20 @@ ideServices.service('RenderSerive',['ResourceService','Upload','$http',function 
     renderer.prototype.renderSlide = function (widget,srcRootDir,dstDir,imgUrlPrefix,cb) {
         var info = widget.info;
         if (!!info){
+            //font
+            var text = '';
+            var style = {};
+            var font = {};
+            font['font-style'] = widget.info.fontItalic;
+            font['font-weight'] = widget.info.fontBold;
+            font['font-size'] = widget.info.fontSize;
+            font['font-family'] = widget.info.fontFamily;
+            font['font-color'] = widget.info.fontColor;
+            style.color = font['font-color'];
+            style.font = (font['font-style']||'')+' '+(font['font-variant']||'')+' '+(font['font-weight']||'')+' '+(font['font-size']||24)+'px'+' '+(font['font-family']||'arial');
+            style.textAlign = 'center';
+            style.textBaseline = 'middle';
+
             //trans each slide
             var width = info.width;
             var height = info.height;
@@ -538,6 +552,13 @@ ideServices.service('RenderSerive',['ResourceService','Upload','$http',function 
                     }
                     renderingX.renderImage(ctx,new Size(width,height),new Pos(),targetImageObj,new Pos(),new Size(width,height));
                 }
+
+                //render font --20170705
+                text = curSlice.text;
+                if(!!text){
+                    renderingX.renderText(ctx,new Size(info.width,info.height),new Pos(),text,style,true,null,this.customFonts);
+                }
+
                 //output
                 var imgName = widget.id.split('.').join('');
                 var outputFilename = imgName +'-'+ i+'.png';
@@ -568,6 +589,73 @@ ideServices.service('RenderSerive',['ResourceService','Upload','$http',function 
 
     };
 
+    renderer.prototype.renderTexNum = function(widget,srcRootDir,dstDir,imgUrlPrefix,cb){
+        var info = widget.info;
+        if (!!info){
+            //trans each slide
+            var width = info.characterW;
+            var height = info.characterH;
+
+            var slideTex = widget.texList[0];
+            var totalSlices = slideTex.slices.length;
+            slideTex.slices.map(function (slice,i) {
+                var canvas = new Canvas(width,height);
+                var ctx = canvas.getContext('2d');
+                var curSlice = slideTex.slices[i];
+                // console.log('slice: ',i,' canas ',canvas,' slice: ',curSlice,width,height);
+                ctx.clearRect(0,0,width,height);
+                ctx.save();
+                //render color
+                renderingX.renderColor(ctx,new Size(width,height),new Pos(),curSlice.color);
+                //render image;
+                var imgSrc = curSlice.imgSrc;
+                if (imgSrc!==''){
+                    var imgUrl = path.join(srcRootDir,imgSrc);
+                    var targetImageObj = this.getTargetImage(imgUrl);
+                    if (!targetImageObj){
+                        //not added to images
+                        var imgObj = new Image();
+                        try{
+                            imgObj.src = loadImageSync(imgUrl);
+                            this.addImage(imgUrl,imgObj);
+                            targetImageObj = imgObj;
+                        }catch (err){
+                            targetImageObj = null;
+                        }
+
+                    }
+                    renderingX.renderImage(ctx,new Size(width,height),new Pos(),targetImageObj,new Pos(),new Size(width,height));
+                }
+
+                //output
+                var imgName = widget.id.split('.').join('');
+                var outputFilename = imgName +'-'+ i+'.png';
+                var outpath = path.join(dstDir,outputFilename);
+                canvas.output(outpath,function (err) {
+                    if (err){
+                        cb && cb(err);
+                    }else{
+                        this.trackedRes.push(new ResTrack(imgSrc,curSlice.color,null,outputFilename,width,height,curSlice));
+                        // console.log(_.cloneDeep(this.trackedRes))
+                        //write widget
+                        curSlice.originSrc = curSlice.imgSrc;
+                        curSlice.imgSrc = path.join(imgUrlPrefix||'',outputFilename);
+                        console.log('keke',outputFilename);
+                        //if last trigger cb
+                        totalSlices -= 1;
+                        if (totalSlices<=0){
+                            cb && cb();
+                        }
+                    }
+                }.bind(this));
+
+                ctx.restore();
+            }.bind(this));
+
+        }else{
+            cb&&cb();
+        }
+    };
 
     renderer.prototype.renderRotateImg = function (widget,srcRootDir,dstDir,imgUrlPrefix,cb) {
         var info = widget.info;
@@ -765,6 +853,7 @@ ideServices.service('RenderSerive',['ResourceService','Upload','$http',function 
     renderer.prototype.renderWidget = function (widget,srcRootDir,dstDir,imgUrlPrefix,cb) {
         switch (widget.subType){
             case 'MyButton':
+            case 'MySwitch':
                 this.renderButton(widget,srcRootDir,dstDir,imgUrlPrefix,cb);
                 break;
             case 'MyButtonGroup':
@@ -786,7 +875,10 @@ ideServices.service('RenderSerive',['ResourceService','Upload','$http',function 
                 this.renderDashboard(widget,srcRootDir,dstDir,imgUrlPrefix,cb);
                 break;
             case 'MyRotateImg':
-                this.renderRotateImg(widget,srcRootDir,dstDir,imgUrlPrefix,cb)
+                this.renderRotateImg(widget,srcRootDir,dstDir,imgUrlPrefix,cb);
+                break;
+            case 'MyTexNum':
+                this.renderTexNum(widget,srcRootDir,dstDir,imgUrlPrefix,cb);
                 break;
             default:
                 cb&&cb();